bc: entproto: convert field.Int into int64 proto fields (#254)

+++ entproto/internal/todo/ent/proto/entpb/entpb_attachment_service.go
@@ -38,13 +38,13 @@ func toProtoAttachment(e *ent.Attachment) (*Attachment, error) {
 	}
 	v.Id = id
 	for _, edg := range e.Edges.Recipients {
-		id := int32(edg.ID)
+		id := int64(edg.ID)
 		v.Recipients = append(v.Recipients, &User{
 			Id: id,
 		})
 	}
 	if edg := e.Edges.User; edg != nil {
-		id := int32(edg.ID)
+		id := int64(edg.ID)
 		v.User = &User{
 			Id: id,
 		}
